Change router scheme to always echo
Some logic is hard to prove right. This patch changes router.c to 'be stupider' instead. Any time a data interrupt is taken, we do all possible things that could be related. For PC_INT0, we echo the rPi outward and the current dispenser back toward the rPi.

Note, even if the current dispenser is tied to INT1, when INT0 fires we still read the current state of our target dispenser line, and set the outbound to the same level. This will cause extra work (due to extraneous interrupts), but is shorter and easier to argue correct.